updated documentation of ECL cluster detection region following suggestions by Sam and Torben, reverted data type to double

+++ analysis/VariableManager/include/Variables.h
@@ -539,8 +539,9 @@ namespace Belle2 {
     /**
      * return 1/2/3 if the ECL Cluster is detected in the forward/barrel/backward region
      * return 11 and 13 if the ECL Cluster is in gap between barrel and forward respectively backward
+     * return 0 otherwise
      */
-    int eclClusterDetectionRegion(const Particle* particle);
+    double eclClusterDetectionRegion(const Particle* particle);
 
     /**
      * return distance from eclCluster to nearest track hitting the ECLCluster
